webapp/requirements.txt
python-dotenv
azure-storage-blob
azure-cosmosdb-table
Flask
